feat(web): move templates to settings page
- Extracted template management logic into `TemplateManager` component.
- Added 'Templates' tab to the Settings page.
- Removed standalone templates page and navbar link.
- Integrated `TemplateManager` into the new Settings tab.
